Fix layout overlap, int overflow in delay loops, stale editor race, and update docs (#854, #855, #856, #882)
- Compute text-based size overrides for all element types (stocks, modules,
  AUX, lookup, CLD) before running AutoLayout so the layout engine accounts
  for actual rendered dimensions and avoids element overlap
- Change catch-up loop counter from int to long in Delay1, Delay3, Trend,
  Forecast, and Npv to prevent overflow when delta exceeds Integer.MAX_VALUE
- Capture editor reference before spawning async layout thread in
  ModelWindow.loadDefinition() to prevent stale editor/view mismatch on
  rapid file opens
- Update README, Quickstart, and Visual Editor Guide to reflect 100+ bundled
  models across 18 categories and replace production disclaimer with release
  label

+++ courant-engine/src/main/java/systems/courant/sd/model/Delay3.java
@@ -133,7 +133,7 @@ public class Delay3 implements Formula, Resettable {
             double stageTime = delayTime / 3.0;
             long delta = step - lastStep;
             double inputVal = input.getAsDouble();
-            for (int d = 0; d < delta; d++) {
+            for (long d = 0; d < delta; d++) {
                 // Compute outflow rates from current stage levels
                 double rate1 = stage1 / stageTime;
                 double rate2 = stage2 / stageTime;
